Initialize internationalization based on locale
for now the preferred locale should be changed in the url. This will be saved in the cookies, so that next time the user faces a good user experience, as their saved locale preference will be used.

A button still should be implemented to toggle between languages.

Test are as well to be implemented.

WIP #18